Reconstruction of the library according to Angular CLI generation.
