--- .npmignore
@@ -1,4 +1,0 @@
-.git
-.idea
-.eslintrc
-.travis.yml
