--- .nycrc.json
@@ -1,12 +1,0 @@
-{
-  "all": true,
-  "check-coverage": true,
-  "report-dir": "./coverage",
-  "reporter": [
-    "lcov",
-    "text"
-  ],
-  "include": [
-    "src/**/*.ts"
-  ]
-}
