+++ .travis.yml
@@ -8,6 +8,7 @@ os:
   - windows
 sudo: false
 node_js:
+  - 15
   - 14
   - 13
   - 12
