--- jasmine.json
@@ -1,8 +1,0 @@
-{
-  "spec_dir": "test",
-  "spec_files": [
-    "ng-event-bus.spec.ts"
-  ],
-  "stopSpecOnExpectationFailure": false,
-  "random": true
-}
