+++ ng-package.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "./node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "./dist",
+  "whitelistedNonPeerDependencies": [
+    "rxjs"
+  ],
+  "lib": {
+    "entryFile": "src/public-api.ts"
+  }
+}
