+++ package.json
@@ -7,12 +7,8 @@
     "email": "contact@cristiammercado.com",
     "url": "https://github.com/cristiammercado/ng-event-bus/issues"
   },
-  "main": "dist/ng_event_bus.js",
-  "types": "dist/ng_event_bus.d.ts",
-  "scripts": {
-    "build": "./node_modules/.bin/tsc -d ./src/ng_event_bus.ts --lib es6 --outDir ./dist",
-    "test": "nyc jasmine-ts --config=jasmine.json"
-  },
+  "main": "dist/ng-event-bus.js",
+  "types": "dist/ng-event-bus.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/cristiammercado/ng-event-bus.git"
@@ -31,20 +27,39 @@
     "url": "https://www.cristiammercado.com/en"
   },
   "license": "MIT",
+  "private": true,
+  "scripts": {
+    "ng": "ng",
+    "build": "ng build --prod",
+    "test": "ng test --watch=false --codeCoverage",
+    "lint": "ng lint",
+    "prettier:write": "prettier -w \"./src/**/*.ts\"",
+    "prettier:check": "prettier -l \"./src/**/*.ts\""
+  },
   "dependencies": {
-    "rxjs": "^6.6.3"
+    "rxjs": "~6.6.3"
   },
   "devDependencies": {
-    "@types/jasmine": "^3.5.14",
-    "jasmine": "^3.6.2",
-    "jasmine-core": "^3.6.0",
-    "jasmine-ts": "^0.3.0",
-    "mem": "^8.0.0",
-    "minimist": "^1.2.5",
-    "nyc": "^15.1.0",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7",
-    "uuid": "^8.3.1"
+    "@angular-devkit/build-angular": "~0.1100.2",
+    "@angular/cli": "~11.0.2",
+    "@angular/compiler-cli": "~11.0.2",
+    "@types/jasmine": "~3.6.2",
+    "@types/node": "~14.14.9",
+    "codelyzer": "~6.0.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "~5.2.3",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "~1.5.4",
+    "ng-packagr": "~11.0.2",
+    "prettier": "~2.2.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~9.0.0",
+    "tslib": "~2.0.3",
+    "tslint": "~6.1.3",
+    "typescript": "~4.0.5"
   },
   "engines": {
     "node": ">= 10"
