+++ src/public-api.ts
@@ -1,0 +1,4 @@
+/*
+ * Public API Surface of ng-event-bus
+ */
+export * from './lib/ng-event-bus';
