+++ tsconfig.spec.json
@@ -1,0 +1,16 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "dist/out-tsc/spec",
+    "types": [
+      "jasmine"
+    ]
+  },
+  "files": [
+    "./src/test.ts"
+  ],
+  "include": [
+    "./src/**/*.spec.ts",
+    "./src/**/*.d.ts"
+  ]
+}
